WIP: Linux port  - Cursor clipping now clips the cursor instead of wrapping it  - Drop target now working properly
Former-commit-id: 932c3ee8316004c8decb503b871a7d628c9e810b

+++ Source/SBansheeEditor/Wrappers/BsScriptSceneSelection.cpp
@@ -126,10 +126,10 @@ namespace bs
 			{
 				Vector<HSceneObject> selectedSOs = Selection::instance().getSceneObjects();
 
-				for (int i = 0; i < pickedObjects.size(); i++) 
+				for (int i = 0; i < (int)pickedObjects.size(); i++)
 				{
 					bool found = false;
-					for (int j = 0; j < selectedSOs.size(); j++)
+					for (int j = 0; j < (int)selectedSOs.size(); j++)
 					{
 						if (selectedSOs[j] == pickedObjects[i])
 						{
